third_party/WebKit/LayoutTests/crypto/aes-gcm-encrypt-decrypt-expected.txt

Tests encrypt/decrypt for AES-GCM

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "AES-GCM"
PASS tmpKey.usages.join(',') is "encrypt,decrypt"
PASS: Encryption should be [72ac8493e3a5228b5d130a69d2510e42] and was
PASS: Decryption should be [] and was
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "AES-GCM"
PASS tmpKey.usages.join(',') is "encrypt,decrypt"
PASS: Encryption should be [751f3098d59cf4ea1d2fb0853bde1c] and was
PASS: Decryption should be [] and was
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "AES-GCM"
PASS tmpKey.usages.join(',') is "encrypt,decrypt"
PASS: Encryption should be [bdd2ed6c66fa087dce617d7fd1ff6d93ba82e49c55a22ed02ca67da4ec6f] and was
PASS: Decryption should be [bfaaaea3880d72d4378561e2597a9b35] and was
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "AES-GCM"
PASS tmpKey.usages.join(',') is "encrypt,decrypt"
PASS: Encryption should be [c4489fa64a6edf80e7e6a3b8855bc37c772ee7de0f91a981c36c93a35c88] and was
PASS: Decryption should be [d30b02c343487105219d6fa080acc743] and was
PASS successfullyParsed is true

TEST COMPLETE
